Tied MISO to 1, so it should only send 1s.

+++ armsrc/fpgaloader.c
@@ -527,7 +527,7 @@ void FpgaSendCommand(uint16_t cmd, uint16_t v)
 	Dbprintf("Just sent an FPGA command over SPI. Value of RDRF (Receive Data Register Full) = %d", (AT91C_BASE_SPI->SPI_SR & AT91C_SPI_RDRF));
 	uint16_t recvd_from_spi;
 	if((AT91C_BASE_SPI->SPI_SR & AT91C_SPI_RDRF)){
-		recvd_from_spi = AT91C_BASE_SPI->SPI_RDR;
+		recvd_from_spi = (uint16_t)AT91C_BASE_SPI->SPI_RDR;
 		Dbprintf("Just read the value. Value = %x. Receive data register full = %x", recvd_from_spi, (AT91C_BASE_SPI->SPI_SR & AT91C_SPI_RDRF));
 	}
 }
